move to micronaut-langchain-store-neo4j
