+++ gradle/libs.versions.toml
@@ -96,7 +96,6 @@ langchain4j-elasticsearch = { module = 'dev.langchain4j:langchain4j-elasticsearc
 langchain4j-pgvector = { module = 'dev.langchain4j:langchain4j-pgvector' }
 langchain4j-qdrant = { module = 'dev.langchain4j:langchain4j-qdrant' }
 langchain4j-google-ai-gemini = { module = 'dev.langchain4j:langchain4j-google-ai-gemini' }
-langchain4j-community-neo4j = { module = 'dev.langchain4j:langchain4j-community-neo4j' }
 langchain4j-cassandra = { module = 'dev.langchain4j:langchain4j-cassandra' }
 langchain4j-embeddings-all-minilm-l6-v2 = { module = 'dev.langchain4j:langchain4j-embeddings-all-minilm-l6-v2' }
 junit-platform-engine = { module = "org.junit.platform:junit-platform-suite-engine" }
