+++ test-suite/build.gradle.kts
@@ -12,11 +12,9 @@ dependencies {
     testImplementation(project(":micronaut-langchain4j-ollama"))
 
     testImplementation(project(":test-suite-utils"))
-    testImplementation(platform(mnTestResources.boms.testcontainers))
-    testImplementation(platform(libs.langchain4j.community.bom))
     testImplementation(project(":micronaut-langchain4j-store-redis"))
     testAnnotationProcessor(mn.micronaut.inject.java)
-    testImplementation(libs.langchain4j.community.neo4j)
+    testImplementation(project(":micronaut-langchain4j-store-neo4j"))
     testImplementation(libs.langchain4j.cassandra)
     testImplementation(mnTest.micronaut.test.junit5)
     testImplementation(mnTest.junit.jupiter.engine)
